requirements.txt
gradio
google-generativeai
sentence-transformers
faiss-cpu
tqdm
numpy
pandas
python-dotenv